Hide save button for read-only Scribbles.

+++ app/View/Elements/scribble-input-form.ctp
@@ -6,6 +6,8 @@ echo $this->element("latex-selector");
 echo $this->Form->input("body", array("rows" => "10", "id" => "ScribbleInput", "class" => "form-control", "style" => "", "div" => array("class" => "form-group")));
 if (isset($edit) && $edit === true)
 	echo $this->Form->input("id", array("type" => "hidden"));
-echo $this->Form->submit("Scribble!", array("class" => "btn btn-default"));
+if (!isset($scribble) || !$scribble["Scribble"]["read_only"]) {
+	echo $this->Form->submit("Scribble!", array("class" => "btn btn-default"));
+}
 echo $this->Form->end();
 ?>
